Work-around for compiler crash.
Kernel argument triggering the crash is of type StepInfoAll. Remove
argument from kernel invocation and declaration/definition, and declare
it as a local variable (uninitialized) in the kernel definition. Purpose
is to get build process continuing for now. If this works, then we can
replace the pass-by-value argument by a device allocation and a host-device
copy.