Remove __restrict__
Compiling hipified code where __restrict__ pointers are passed to
kernels with corresponding __restrict__ pointer arguments is broken.
In intermediate layers, the pointers are not __restrict__, which
results in a T* __restrict__ being initialized with a T*, which is
not allowed.

This commit replaces all '__restrict__' by 'RESTRICT    ' (same string length
as to not mess up compilation), and #defines RESTRICT to be empty in
arguments.h.

This is a quick hack to get compilation going. Better solution:
1. Fix the handling of __restrict__ pointers in HIP/hcc
2. Pending 1, instead of removing all occurrences of __restrict__, remove
   only the ones that are really needed
3. Check the appropriate location for '#define RESTRICT', and wrap it
   in a conditional, e.g., '#ifdef USE_HIP'